API/requirements.txt
gunicorn==19.9.0
Flask==0.12.2
Flask-Cors==3.0.3
tensorflow==1.11.0
Keras==2.2.4
numpy==1.15.1
Pillow==5.0.0
